Move theme switcher to own component
